[step3] Cars에게 일을 부여
